idea_factory/requirements.txt
Flask==3.0.0
openai>=1.0.0
google-api-python-client==2.149.0
google-auth==2.35.0
Flask-Limiter==3.8.0
python-dotenv==1.0.0
requests==2.31.0

